refactor : 동적 Html 랜더링 관련 클래스 패키지 변경
